refactor(Sources): Move CLI commands back into `stencila` crate because it requires`PROJECTS`
