+++ rust/stencila/src/lib.rs
@@ -5,11 +5,11 @@
 //
 // Core Stencila objects e.g `File`, `Article`, `Project`
 
-pub mod conversions;
 pub mod documents;
 pub use kernels;
 pub mod projects;
 pub mod sessions;
+pub mod sources;
 
 // Features
 //
